+++ lib/karafka/pro/processing/strategies/aj/dlq_lrj_mom_vp.rb
@@ -51,14 +51,12 @@ module Karafka
                   seek(coordinator.seek_offset, false) unless revoked?
 
                   resume
-                elsif coordinator.pause_tracker.attempt <= topic.dead_letter_queue.max_retries
-                  retry_after_pause
                 else
-                  coordinator.pause_tracker.reset
-                  skippable_message, = find_skippable_message
-                  dispatch_to_dlq(skippable_message) if dispatch_to_dlq?
-                  mark_as_consumed(skippable_message)
-                  pause(coordinator.seek_offset, nil, false)
+                  apply_dlq_flow do
+                    skippable_message, = find_skippable_message
+                    dispatch_to_dlq(skippable_message) if dispatch_to_dlq?
+                    mark_as_consumed(skippable_message)
+                  end
                 end
               end
             end