+++ lib/karafka/pro/processing/strategies/aj/dlq_mom_vp.rb
@@ -48,16 +48,14 @@ module Karafka
                   return if revoked?
 
                   mark_as_consumed(last_group_message)
-                elsif coordinator.pause_tracker.attempt <= topic.dead_letter_queue.max_retries
-                  retry_after_pause
                 else
-                  # Here we are in a collapsed state, hence we can apply the same logic as
-                  # Aj::DlqMom
-                  coordinator.pause_tracker.reset
-                  skippable_message, = find_skippable_message
-                  dispatch_to_dlq(skippable_message) if dispatch_to_dlq?
-                  mark_as_consumed(skippable_message)
-                  pause(coordinator.seek_offset, nil, false)
+                  apply_dlq_flow do
+                    # Here we are in a collapsed state, hence we can apply the same logic as
+                    # Aj::DlqMom
+                    skippable_message, = find_skippable_message
+                    dispatch_to_dlq(skippable_message) if dispatch_to_dlq?
+                    mark_as_consumed(skippable_message)
+                  end
                 end
               end
             end