+++ lib/karafka/pro/processing/strategies/dlq/mom.rb
@@ -35,28 +35,18 @@ module Karafka
 
                 if coordinator.success?
                   coordinator.pause_tracker.reset
-                elsif coordinator.pause_tracker.attempt <= topic.dead_letter_queue.max_retries
-                  retry_after_pause
-                # If we've reached number of retries that we could, we need to skip the first
-                # message that was not marked as consumed, pause and continue, while also moving
-                # this message to the dead topic.
-                #
-                # For a Mom setup, this means, that user has to manage the checkpointing by
-                # himself. If no checkpointing is ever done, we end up with an endless loop.
                 else
-                  # We reset the pause to indicate we will now consider it as "ok".
-                  coordinator.pause_tracker.reset
+                  apply_dlq_flow do
+                    skippable_message, = find_skippable_message
+                    dispatch_to_dlq(skippable_message) if dispatch_to_dlq?
 
-                  skippable_message, = find_skippable_message
-                  dispatch_to_dlq(skippable_message) if dispatch_to_dlq?
-
-                  # Save the next offset we want to go with after moving given message to DLQ
-                  # Without this, we would not be able to move forward and we would end up
-                  # in an infinite loop trying to un-pause from the message we've already processed
-                  # Of course, since it's a MoM a rebalance or kill, will move it back as no
-                  # offsets are being committed
-                  coordinator.seek_offset = skippable_message.offset + 1
-                  pause(coordinator.seek_offset, nil, false)
+                    # Save the next offset we want to go with after moving given message to DLQ
+                    # Without this, we would not be able to move forward and we would end up
+                    # in an infinite loop trying to un-pause from the message we've already
+                    # processed. Of course, since it's a MoM a rebalance or kill, will move it back
+                    # as no offsets are being committed
+                    coordinator.seek_offset = skippable_message.offset + 1
+                  end
                 end
               end
             end