+++ spec/lib/karafka/pro/routing/features/dead_letter_queue/contracts/topic_spec.rb
@@ -7,7 +7,8 @@ RSpec.describe_current do
     {
       dead_letter_queue: {
         active: true,
-        max_retries: 1
+        max_retries: 1,
+        strategy: -> {}
       },
       virtual_partitions: {
         active: false
@@ -17,6 +18,12 @@ RSpec.describe_current do
 
   context 'when config is valid' do
     it { expect(check).to be_success }
+  end
+
+  context 'when strategy is not callable' do
+    before { config[:dead_letter_queue][:strategy] = 1 }
+
+    it { expect(check).not_to be_success }
   end
 
   context 'when trying to use DLQ with VP without any retries' do