+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Karafka framework changelog
 
 ## 2.3.2 (Unreleased)
+- [Enhancement] Provide `Consumer#errors_tracker` to be able to get errors that occurred while doing complex recovery.
 - [Fix] Quieting mode causes too early unsubscribe.
 
 ## 2.3.1 (2024-02-08)